requirements.txt
streamlit
yfinance
nltk
pandas
numpy<2
scikit-learn
rank_bm25
sentence-transformers
transformers
torch==2.2.0
setuptools>=65.5.0
faiss-cpu
os
sentencepiece